requirements.txt
# Core dependencies
requests>=2.28.0
beautifulsoup4>=4.9.3
PyYAML>=5.4.1
aiohttp>=3.8.1
jinja2>=3.0.1

# Authentication and security
python-jwt>=3.3.0
paramiko>=2.8.1

# Network scanning
python-nmap>=0.7.1

# Reporting
reportlab>=3.6.8
matplotlib>=3.4.3

# Development
setuptools>=65.0.0
flask>=2.0.0
pytest>=7.0.0
black>=22.0.0
mypy>=0.950